requirements.txt
boto3>=1.34.80
aws-cdk-lib==2.136.0
constructs>=10.0.0,<11.0.0
